to use config files for ga, #7, #1
adds the use of config files to the genetic algorithm. A new CLI is added, which takes a
filepath to a config file, which specifies the configuration for the genetic algorithm.
That configuration will be read by the config module, saved and other modules get acces
to these global variables, which won't be changed after the initilization through the
configuration file.

The prototype version has been tested a bit.

The RaLaNS version hasn't been tested yet.