corrects rare case if select is too high
if all are select, then there are not enough remaining individuals to sample
from, if that is the case, then just the remaining number of individuals will
be chosen and not the number, which is specified by the tournsize.
So f.x. tournsize is 3, but there are just 2 individuals left.